InlineEdit: Provide setting to turn it off
When edit workflow is disabled, the following features are turned off:

* Create a empty change from the UI or REST API
* Create a follow-up change from the UI or REST API
* Push a new edit through REST API:
** amend commit message
** change file
** restore deleted file
** delete file
* Push a new edit through magic branch option
* Publish change edit
* Delete change edit
* Rebase change edit
* Switch to edit mode on the change screen
* Edit content of a file in the UI
* Edit icon in patch set select box
* Fix button in publish draft box
* Key binding "e" to edit file content in side-by-side screen
* Edits entry in "My" menu

Feature: Issue 3596

+++ gerrit-server/src/main/java/com/google/gerrit/server/change/DeleteChangeEdit.java
@@ -16,14 +16,18 @@ package com.google.gerrit.server.change;
 
 import com.google.common.base.Optional;
 import com.google.gerrit.extensions.restapi.AuthException;
+import com.google.gerrit.extensions.restapi.BadRequestException;
 import com.google.gerrit.extensions.restapi.ResourceNotFoundException;
 import com.google.gerrit.extensions.restapi.Response;
 import com.google.gerrit.extensions.restapi.RestModifyView;
 import com.google.gerrit.server.change.DeleteChangeEdit.Input;
+import com.google.gerrit.server.config.GerritServerConfig;
 import com.google.gerrit.server.edit.ChangeEdit;
 import com.google.gerrit.server.edit.ChangeEditUtil;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+
+import org.eclipse.jgit.lib.Config;
 
 import java.io.IOException;
 
@@ -33,15 +37,22 @@ public class DeleteChangeEdit implements RestModifyView<ChangeResource, Input> {
   }
 
   private final ChangeEditUtil editUtil;
+  private final boolean allowEdits;
 
   @Inject
-  DeleteChangeEdit(ChangeEditUtil editUtil) {
+  DeleteChangeEdit(@GerritServerConfig Config cfg,
+      ChangeEditUtil editUtil) {
     this.editUtil = editUtil;
+    allowEdits = cfg.getBoolean("change", "allowEdits", true);
   }
 
   @Override
   public Response<?> apply(ChangeResource rsrc, Input input)
-      throws AuthException, ResourceNotFoundException, IOException {
+      throws AuthException, ResourceNotFoundException, IOException,
+      BadRequestException {
+    if (!allowEdits) {
+      throw new BadRequestException("edit workflow is disabled");
+    }
     Optional<ChangeEdit> edit = editUtil.byChange(rsrc.getChange());
     if (edit.isPresent()) {
       editUtil.delete(edit.get());
